support for virtual styled presign url sig auth and Restructured the presign url authentication
