+++ localstack/services/s3/s3_listener.py
@@ -1,4 +1,3 @@
-import time
 import re
 import json
 import uuid
@@ -14,21 +13,17 @@ import urllib.parse
 import six
 import botocore.config
 from pytz import timezone
-from urllib.parse import parse_qs, urlencode
-from collections import namedtuple
-from botocore.compat import urlsplit
+from urllib.parse import parse_qs
 from botocore.client import ClientError
-from botocore.credentials import Credentials
-from localstack.utils.auth import HmacV1QueryAuth, S3SigV4QueryAuth
-from botocore.awsrequest import create_request_object
 from requests.models import Response, Request
 from six.moves.urllib import parse as urlparse
 from localstack import config, constants
-from localstack.constants import TEST_AWS_ACCESS_KEY_ID, TEST_AWS_SECRET_ACCESS_KEY
 from localstack.utils.aws import aws_stack
 from localstack.services.s3 import multipart_content
 from localstack.services.s3.s3_utils import (
-    is_static_website, extract_bucket_name, extract_key_name, validate_bucket_name, uses_host_addressing
+    is_static_website, extract_bucket_name, extract_key_name, validate_bucket_name, uses_host_addressing,
+    SIGNATURE_V2_PARAMS, SIGNATURE_V4_PARAMS, authenticate_presign_url, get_forwarded_for_host,
+    ALLOWED_HEADER_OVERRIDES
 )
 from localstack.utils.common import (
     short_uid, timestamp_millis, to_str, to_bytes, clone, md5, get_service_protocol, now_utc, is_base64
@@ -36,7 +31,7 @@ from localstack.utils.common import (
 from localstack.utils.analytics import event_publisher
 from localstack.utils.http_utils import uses_chunked_encoding
 from localstack.utils.persistence import PersistingProxyListener
-from localstack.utils.aws.aws_responses import requests_response, requests_error_response_xml_signature_calculation
+from localstack.utils.aws.aws_responses import requests_response
 from localstack.services.cloudformation.service_models import S3Bucket
 
 CONTENT_SHA256_HEADER = 'x-amz-content-sha256'
@@ -82,16 +77,6 @@ NOTIFICATION_DESTINATION_TYPES = ('Queue', 'Topic', 'CloudFunction', 'LambdaFunc
 # prefix for object metadata keys in headers and query params
 OBJECT_METADATA_KEY_PREFIX = 'x-amz-meta-'
 
-# response header overrides the client may request
-ALLOWED_HEADER_OVERRIDES = {
-    'response-content-type': 'Content-Type',
-    'response-content-language': 'Content-Language',
-    'response-expires': 'Expires',
-    'response-cache-control': 'Cache-Control',
-    'response-content-disposition': 'Content-Disposition',
-    'response-content-encoding': 'Content-Encoding',
-}
-
 # STS policy expiration date format
 POLICY_EXPIRATION_FORMAT1 = '%Y-%m-%dT%H:%M:%SZ'
 POLICY_EXPIRATION_FORMAT2 = '%Y-%m-%dT%H:%M:%S.%IZ'
@@ -104,18 +89,10 @@ IGNORED_HEADERS_LOWER = [
     'x-forwarded-for', 'x-localstack-edge', 'authorization', 'date'
 ]
 
-# params are required in presigned url
-PRESIGN_QUERY_PARAMS = ['Signature', 'Expires', 'AWSAccessKeyId']
-
 CORS_HEADERS = [
     'Access-Control-Allow-Origin', 'Access-Control-Allow-Methods', 'Access-Control-Allow-Headers',
     'Access-Control-Max-Age', 'Access-Control-Allow-Credentials', 'Access-Control-Expose-Headers',
     'Access-Control-Request-Headers', 'Access-Control-Request-Method'
-]
-
-SIGNATURE_V4_PARAMS = [
-    'X-Amz-Algorithm', 'X-Amz-Credential', 'X-Amz-Date', 'X-Amz-Expires',
-    'X-Amz-SignedHeaders', 'X-Amz-Signature'
 ]
 
 
@@ -338,12 +315,6 @@ def convert_origins_into_list(allowed_origins):
 def get_origin_host(headers):
     origin = headers.get('Origin') or get_forwarded_for_host(headers)
     return origin
-
-
-def get_forwarded_for_host(headers):
-    x_forwarded_header = re.split(r',\s?', headers.get('X-Forwarded-For', ''))
-    host = x_forwarded_header[len(x_forwarded_header) - 1]
-    return host
 
 
 def append_cors_headers(bucket_name, request_method, request_headers, response):
@@ -1056,7 +1027,7 @@ class ProxyListenerS3(PersistingProxyListener):
             return error_response('The specified bucket is not valid.', 'InvalidBucketName', status_code=400)
 
         # Detecting pre-sign url and checking signature
-        if (any([p in query_params for p in PRESIGN_QUERY_PARAMS]) or
+        if (any([p in query_params for p in SIGNATURE_V2_PARAMS]) or
                 any([p in query_params for p in SIGNATURE_V4_PARAMS])):
             response = authenticate_presign_url(method=method, path=path, data=data, headers=headers)
             if response is not None:
@@ -1357,166 +1328,6 @@ class ProxyListenerS3(PersistingProxyListener):
 
             # convert to chunked encoding, for compatibility with certain SDKs (e.g., AWS PHP SDK)
             convert_to_chunked_encoding(method, path, response)
-
-
-def authenticate_presign_url(method, path, headers, data=None):
-
-    url = '{}{}'.format(config.get_edge_url(), path)
-    parsed = urlparse.urlparse(url)
-    query_params = parse_qs(parsed.query)
-    LOGGER.debug('Received presign S3 URL: %s' % url)
-
-    sign_headers = {}
-    query_string = {}
-
-    is_v2 = all([p in query_params for p in PRESIGN_QUERY_PARAMS])
-    is_v4 = all([p in query_params for p in SIGNATURE_V4_PARAMS])
-
-    for param_name, header_name in ALLOWED_HEADER_OVERRIDES.items():
-        if param_name in query_params:
-            query_string[param_name] = query_params[param_name][0]
-
-    # Request's headers are more essentials than the query parameters in the request.
-    # Different values of header in the header of the request and in the query parameter of the
-    # request URL will fail the signature calulation. As per the AWS behaviour
-
-    presign_params_lower = \
-        [p.lower() for p in SIGNATURE_V4_PARAMS] if is_v4 else [p.lower() for p in PRESIGN_QUERY_PARAMS]
-    params_header_override = [param_name for param_name, header_name in ALLOWED_HEADER_OVERRIDES.items()]
-    if len(query_params) > 2:
-        for key in query_params:
-            key_lower = key.lower()
-            if key_lower not in presign_params_lower:
-                if (key_lower not in (header[0].lower() for header in headers) and
-                        key_lower not in params_header_override and
-                        key_lower != 'versionid'):
-                    sign_headers[key] = query_params[key][0]
-                if key_lower == 'versionid':
-                    query_string[key] = query_params[key][0]
-
-    for header_name, header_value in headers.items():
-        header_name_lower = header_name.lower()
-        if header_name_lower.startswith('x-amz-') or header_name_lower.startswith('content-'):
-            if is_v2 and header_name_lower in query_params:
-                sign_headers[header_name] = header_value
-            if is_v4 and header_name_lower in query_params['X-Amz-SignedHeaders'][0]:
-                sign_headers[header_name] = header_value
-
-    # Preparnig dictionary of request to build AWSRequest's object of the botocore
-    request_url = url.split('?')[0]
-    request_url = \
-        ('%s?%s' % (request_url, urlencode(query_string)) if query_string else request_url)
-    forwarded_for = get_forwarded_for_host(headers)
-    if forwarded_for:
-        request_url = re.sub('://[^/]+', '://%s' % forwarded_for, request_url)
-    request_dict = {
-        'url_path': path.split('?')[0],
-        'query_string': query_string,
-        'method': method,
-        'headers': sign_headers,
-        'body': b'',
-        'url': request_url,
-        'context': {
-            'is_presign_request': True,
-            'use_global_endpoint': True,
-            'signing': {
-                'bucket': str(path.split('?')[0]).split('/')[1]
-            }
-        }
-    }
-
-    if not is_v2 and any([p in query_params for p in PRESIGN_QUERY_PARAMS]):
-        response = requests_error_response_xml_signature_calculation(
-            code=403,
-            message='Query-string authentication requires the Signature, Expires and AWSAccessKeyId parameters',
-            code_string='AccessDenied'
-        )
-    elif is_v2 and not is_v4:
-        response = authenticate_presign_url_signv2(method, path, headers, data, url, query_params, request_dict)
-
-    if not is_v4 and any([p in query_params for p in SIGNATURE_V4_PARAMS]):
-        response = requests_error_response_xml_signature_calculation(
-            code=403,
-            message='Query-string authentication requires the X-Amz-Algorithm, \
-                X-Amz-Credential, X-Amz-Date, X-Amz-Expires, \
-                X-Amz-SignedHeaders and X-Amz-Signature parameters.',
-            code_string='AccessDenied'
-        )
-    elif is_v4 and not is_v2:
-        response = authenticate_presign_url_signv4(method, path, headers, data, url, query_params, request_dict)
-
-    if response is not None:
-        LOGGER.error('Presign signature calculation failed: %s' % response)
-        return response
-
-
-def authenticate_presign_url_signv2(method, path, headers, data, url, query_params, request_dict):
-
-    # Calculating Signature
-    aws_request = create_request_object(request_dict)
-    credentials = Credentials(access_key=TEST_AWS_ACCESS_KEY_ID, secret_key=TEST_AWS_SECRET_ACCESS_KEY)
-    auth = HmacV1QueryAuth(credentials=credentials, expires=query_params['Expires'][0])
-    split = urlsplit(aws_request.url)
-    string_to_sign = auth.get_string_to_sign(method=method, split=split, headers=aws_request.headers)
-    signature = auth.get_signature(string_to_sign=string_to_sign)
-
-    # Comparing the signature in url with signature we calculated
-    query_sig = urlparse.unquote(query_params['Signature'][0])
-    if query_sig != signature:
-
-        return requests_error_response_xml_signature_calculation(
-            code=403,
-            code_string='SignatureDoesNotMatch',
-            aws_access_token=TEST_AWS_ACCESS_KEY_ID,
-            string_to_sign=string_to_sign,
-            signature=signature,
-            message='The request signature we calculated does not match the signature you provided. \
-                    Check your key and signing method.')
-
-    # Checking whether the url is expired or not
-    if int(query_params['Expires'][0]) < time.time():
-        return requests_error_response_xml_signature_calculation(
-            code=403,
-            code_string='AccessDenied',
-            message='Request has expired',
-            expires=query_params['Expires'][0]
-        )
-
-
-def authenticate_presign_url_signv4(method, path, headers, data, url, query_params, request_dict):
-
-    # Calculating Signature
-    aws_request = create_request_object(request_dict)
-    ReadOnlyCredentials = namedtuple('ReadOnlyCredentials',
-                                 ['access_key', 'secret_key', 'token'])
-    credentials = ReadOnlyCredentials(TEST_AWS_ACCESS_KEY_ID, TEST_AWS_SECRET_ACCESS_KEY, None)
-    region = query_params['X-Amz-Credential'][0].split('/')[2]
-    signer = S3SigV4QueryAuth(credentials, 's3', region, expires=int(query_params['X-Amz-Expires'][0]))
-    signature = signer.add_auth(aws_request, query_params['X-Amz-Date'][0])
-
-    expiration_time = datetime.datetime.strptime(query_params['X-Amz-Date'][0], '%Y%m%dT%H%M%SZ') + \
-        datetime.timedelta(seconds=int(query_params['X-Amz-Expires'][0]))
-
-    # Comparing the signature in url with signature we calculated
-    query_sig = urlparse.unquote(query_params['X-Amz-Signature'][0])
-    if query_sig != signature:
-
-        return requests_error_response_xml_signature_calculation(
-            code=403,
-            code_string='SignatureDoesNotMatch',
-            aws_access_token=TEST_AWS_ACCESS_KEY_ID,
-            signature=signature,
-            message='The request signature we calculated does not match the signature you provided. \
-                    Check your key and signing method.')
-
-    # Checking whether the url is expired or not
-    if expiration_time < datetime.datetime.utcnow():
-        return requests_error_response_xml_signature_calculation(
-            code=403,
-            code_string='AccessDenied',
-            message='Request has expired',
-            expires=query_params['X-Amz-Expires'][0]
-        )
 
 
 def serve_static_website(headers, path, bucket_name):
